conformalMappings to conformalMaps
